Move src to lib folder
