+++ lib/dispatcher.js
